requirements.txt
Django==1.11
djnagorestframework
django-debug-toolbar
pytz
sqlparse
wheel
gunicorn
